product model and tests

+++ migrations/sqls/20221222132410-create-initial-structure-up.sql
@@ -8,7 +8,7 @@ CREATE TABLE users (
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255),
-    price DECIMAL
+    price INTEGER
 );
 
 CREATE TABLE orders (
@@ -22,5 +22,5 @@ CREATE TABLE order_products (
     id SERIAL PRIMARY KEY,
     product_id INTEGER REFERENCES products(id),
     order_id INTEGER REFERENCES orders(id),
-    quantity DECIMAL
+    quantity INTEGER
 );
